Move files to break circular depedency between test/new-e2e and e2e-framework
