+++ test/e2e-framework/components/windows/command/agent.go
