+++ test/e2e-framework/components/windows/command/product.go
