+++ test/e2e-framework/components/windows/command/registry.go
