+++ test/e2e-framework/components/windows/defender/component.go
@@ -13,9 +13,8 @@ import (
 	"github.com/DataDog/datadog-agent/test/e2e-framework/common/utils"
 	"github.com/DataDog/datadog-agent/test/e2e-framework/components/command"
 	"github.com/DataDog/datadog-agent/test/e2e-framework/components/remote"
+	"github.com/DataDog/datadog-agent/test/e2e-framework/components/windows/powershell"
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
-
-	"github.com/DataDog/datadog-agent/test/new-e2e/tests/windows/common/powershell"
 )
 
 // Manager contains the resources to manage Windows Defender
