+++ test/e2e-framework/components/windows/defender/params.go
