+++ test/e2e-framework/components/windows/fipsmode/component.go
@@ -12,7 +12,7 @@ import (
 	"github.com/DataDog/datadog-agent/test/e2e-framework/common/namer"
 	"github.com/DataDog/datadog-agent/test/e2e-framework/components/command"
 	"github.com/DataDog/datadog-agent/test/e2e-framework/components/remote"
-	"github.com/DataDog/datadog-agent/test/new-e2e/tests/windows/common/powershell"
+	"github.com/DataDog/datadog-agent/test/e2e-framework/components/windows/powershell"
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
 )
