+++ test/e2e-framework/components/windows/fipsmode/params.go
