+++ test/e2e-framework/scenarios/aws/ec2/windows/run.go
@@ -19,9 +19,9 @@ import (
 	"github.com/DataDog/datadog-agent/test/e2e-framework/scenarios/aws/ec2"
 	fakeintakescenario "github.com/DataDog/datadog-agent/test/e2e-framework/scenarios/aws/fakeintake"
 	"github.com/DataDog/datadog-agent/test/e2e-framework/scenarios/outputs"
-	"github.com/DataDog/datadog-agent/test/new-e2e/tests/windows/components/defender"
-	"github.com/DataDog/datadog-agent/test/new-e2e/tests/windows/components/fipsmode"
-	"github.com/DataDog/datadog-agent/test/new-e2e/tests/windows/components/testsigning"
+	"github.com/DataDog/datadog-agent/test/e2e-framework/components/windows/defender"
+	"github.com/DataDog/datadog-agent/test/e2e-framework/components/windows/fipsmode"
+	"github.com/DataDog/datadog-agent/test/e2e-framework/components/windows/testsigning"
 
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
 )
