+++ test/e2e-framework/testing/provisioners/azure/host/windows/params.go
@@ -13,7 +13,7 @@ import (
 	"github.com/DataDog/datadog-agent/test/e2e-framework/scenarios/azure/fakeintake"
 	"github.com/DataDog/datadog-agent/test/e2e-framework/testing/utils/e2e/client/agentclientparams"
 	"github.com/DataDog/datadog-agent/test/e2e-framework/testing/utils/optional"
-	"github.com/DataDog/datadog-agent/test/new-e2e/tests/windows/components/defender"
+	"github.com/DataDog/datadog-agent/test/e2e-framework/components/windows/defender"
 )
 
 // ProvisionerParams is a set of parameters for the Provisioner.
