+++ test/new-e2e/tests/windows/common/defender.go
@@ -10,7 +10,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/DataDog/datadog-agent/test/new-e2e/tests/windows/common/powershell"
+	"github.com/DataDog/datadog-agent/test/e2e-framework/components/windows/powershell"
 
 	"github.com/DataDog/datadog-agent/test/e2e-framework/testing/components"
 )
